fix: inline-switch helper uses arrays for cases and values instead of comma separated strings
Tailwinds parser to find css class names in
templates interprets comma separated
class names as if the comma belonged to
the class name. Hence a lot of classed were
missing in the generated css file. Putting the
class names into an array helps tailwind
to get the correct class names out of the
templates.