FEATURE: employs a loadDynamic template as a workaround for the unusable lookup helper
When including the loadDynamic template the
path to the template that needs to be loaded has
to be defined with the help of the templatePath parameter. Unfortunately the solution isn't as
dynamic as the name might suggest.
Any template
that should be loaded via that template needs to have a hardcoded include inside the template.